dded plot_routines
