+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>pynmd</name>
+	<name>PY_pynmd</name>
 	<comment></comment>
 	<projects>
 	</projects>
