+++ plotting/__init__.py
@@ -3,3 +3,4 @@ from . import octant_plotting
 from . import plot_settings
 from . import colormaps
 from . import vars_param
+#from . import plot_routines
