+++ plotting/vars_param.py
@@ -71,7 +71,7 @@ defs['saltdif']['cmap']  = ps.jetWoGn()
 defs['dens']['var']   = 'dens'
 defs['dens']['vmin']  = 1021
 defs['dens']['vmax']  = 1025
-defs['dens']['label'] = 'density [kg/m3]'
+defs['dens']['label'] = 'density [kg/m \mathrm$^3$]'
 defs['dens']['format']= '%.1g'
 defs['dens']['cmap']  = ps.jetWoGn()
 
@@ -79,28 +79,28 @@ defs['vort']={}
 defs['vort']['var']   = 'vort'
 defs['vort']['vmin']  = -0.002
 defs['vort']['vmax']  =  0.002
-defs['vort']['label'] = 'Vorticity [s-1]'
+defs['vort']['label'] = 'Vorticity [s$^\mathrm{-1}$]'
 defs['vort']['format']= '%.3g'
 defs['vort']['cmap']  = ps.jetWoGn()
 
 defs['div']['var']   = 'div'
 defs['div']['vmin']  = -0.0015
 defs['div']['vmax']  =  0.0015
-defs['div']['label'] = 'Divergence [s-1]'
+defs['div']['label'] = 'Divergence [s$^\mathrm{-1}$]'
 defs['div']['format']= '%.3g'
 defs['div']['cmap']  = ps.jetWoGn()
 
 defs['u']['var']   = 'u'
 defs['u']['vmin']  = -0.5
 defs['u']['vmax']  =  0.5
-defs['u']['label'] = 'U [ms-1]'
+defs['u']['label'] = 'U [ms$^\mathrm{-1}$]'
 defs['u']['format']= '%.2g'
 defs['u']['cmap']  = ps.jetWoGn()
 
 defs['v']['var']   = 'v'
 defs['v']['vmin']  = -0.5
 defs['v']['vmax']  =  0.5
-defs['v']['label'] = 'V [ms-1]'
+defs['v']['label'] = 'V [ms$^\mathrm{-1}$]'
 defs['v']['format']= '%.2g'
 defs['v']['cmap']  = ps.jetWoGn()
 
@@ -108,7 +108,7 @@ defs['w']={}
 defs['w']['var']   = 'w'
 defs['w']['vmin']  = -0.5e-3
 defs['w']['vmax']  =  0.5e-3
-defs['w']['label'] = 'W [ms-1]'
+defs['w']['label'] = 'W [ms$^\mathrm{-1}$]'
 defs['w']['format']= '%.4g'
 defs['w']['cmap']  = ps.jetWoGn()
 
@@ -116,7 +116,7 @@ defs['uv']={}
 defs['uv']['var']   = 'uv'
 defs['uv']['vmin']  =  0.0
 defs['uv']['vmax']  =  1.0
-defs['uv']['label'] = 'Tot. Vel. [ms-1]'
+defs['uv']['label'] = 'Tot. Vel. [ms$^\mathrm{-1}$]'
 defs['uv']['format']= '%.2g'
 defs['uv']['cmap']  = ps.my_cmap
 
@@ -140,7 +140,7 @@ defs['wcap']={}
 defs['wcap']['var']   = 'Dwave'
 defs['wcap']['vmin']  =  0.
 defs['wcap']['vmax']  =  2.0 
-defs['wcap']['label'] = 'Dissp. Wcap [Wat m-2]'    
+defs['wcap']['label'] = 'Dissp. Wcap [Watm$^\mathrm{-2}$]'    
 defs['wcap']['format']= '%.1g'
 defs['wcap']['cmap']  = ps.my_cmap
 
@@ -166,7 +166,7 @@ defs['tke']={}
 defs['tke']['var']   = 'tke'
 defs['tke']['vmin']  = 0e-4
 defs['tke']['vmax']  = 6e-4
-defs['tke']['label'] = 'TKE [m2s-2]'
+defs['tke']['label'] = 'TKE [m$^ \mathrm{2}s$^ \mathrm{-2}]'
 defs['tke']['format']= '%.1g'
 defs['tke']['cmap']  = ps.jetWoGn()
 
@@ -182,7 +182,7 @@ defs['spm']={}
 defs['spm']['var']   = 'SPM'
 defs['spm']['vmin']  =  0.
 defs['spm']['vmax']  =  2e-2
-defs['spm']['label'] = 'SPM [kgm-3] '   
+defs['spm']['label'] = 'SPM [kgm$^ \mathrm{-3}] '   
 defs['spm']['format']= '%.1g'
 defs['spm']['cmap']  = ps.my_cmap
 
@@ -190,7 +190,7 @@ defs['spmdif']={}
 defs['spmdif']['var']   = 'SPM_diff'
 defs['spmdif']['vmin']  =  -0.01
 defs['spmdif']['vmax']  =   0.01
-defs['spmdif']['label'] = 'SPM diff. [kgm-3] '   
+defs['spmdif']['label'] = 'SPM diff. [kgm$^ \mathrm{-3}] '   
 defs['spmdif']['format']= '%.1g'
 defs['spmdif']['cmap']  = ps.jetWoGn()
